Refactor directory to ac.th.kmitl.it.soa.group5 and class name.
