--- ;
@@ -1,37 +1,0 @@
-# Team Convention กลุ่มที่ : 5
-## ใช้ภาษา Kotlin
-<p align="center">
-  <a href="https://kotlinlang.org/">
-    <img alt="kotlin banner" title="kotlin banner" src="https://kotlinlang.org/assets/images/twitter-card/kotlin_800x320.png" width="800">
-  </a>
-</p>
-
-### การเว้นบรรทัด
-- เว้นบรรทัด 1 บรรทัดระหว่างmethod เพื่อแยกให้เป็นสัดส่วน
-```
-fun method1(){
-    }
-
-fun method2(){
-    }
-```
-- ในการตั้งเงื่อนไขเราจะเว้น white space แบบนี้
-```
-if (ch == 'a' || ch == 'e' || ch == 'i'
-    || ch == 'o' || ch == 'u') {
-```
-### การตั้งชื่อ
-- ชื่อ class ชื่อ แพคเกต ขึ้นต้นตัวใหญ่
-```
-package Week1
-
-class Count_vowel {
-}
-```
-- ในการตั้งชื่อทุกอย่างในโปรเจดจะใช้อันเดอร์สกอร์และหลังอันเดอร์สกอร์จะเป็นตัวเล็ก เช่น variable_example
-- ตัวแปรทั้งหมดจะต้องมีความหมาย
-### อื่นๆ
-- ถ้าจะมีคอมเม้นของ method จะคอมเม้นข้างใต้ชื่อ method เพื่อบอกว่า method นั้นทำงานอย่างไร
-- ถ้าต้องมีการ import จะเอาส่วนของ import ไว้ใต้ package
-- ไม่ใส่ semi colon(;)
-- เว้นวรรค 1 ครั้งระหว่าง condition และใน 1 บรรทัดจะมีไม่เกิน 3 condition ถ้าเกินก็เว้นบรรทัดลงไป
