--- etaxinvoice/src/main/java/th/ac/kmitl/it/group5/etaxinvoice/EtaxinvoiceApplication.java
@@ -1,12 +1,0 @@
-package th.ac.kmitl.it.group5.etaxinvoice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EtaxinvoiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EtaxinvoiceApplication.class, args);
-	}
-}
