--- etaxinvoice/src/main/java/th/ac/kmitl/it/group5/etaxinvoice/GreetingETaxController.java
@@ -1,13 +1,0 @@
-package th.ac.kmitl.it.group5.etaxinvoice;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class GreetingETaxController {
-
-    @RequestMapping("/greetingtax")
-    public String greetingtax() {
-        return "greetingtax";
-    }
-}
