--- etaxinvoice/src/main/java/th/ac/kmitl/it/group5/etaxinvoice/IndexController.java
@@ -1,13 +1,0 @@
-package th.ac.kmitl.it.group5.etaxinvoice;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class IndexController {
-
-    @RequestMapping("/")
-    public String index() {
-        return "index";
-    }
-}
