--- etaxinvoice/src/main/resources/templates/greetingtax.html
@@ -1,10 +1,0 @@
-<!DOCTYPE HTML>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>Greeting your tax</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-</head>
-<body>
-    <p th:text="'Hello, Tax!'" />
-</body>
-</html>
