--- etaxinvoice/src/main/resources/templates/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-<head>
-    <title>Getting Started: Serving Web Content</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-</head>
-<body>
-    <p>Get your greeting tax <a href="/greetingtax">here</a></p>
-</body>
-</html>
