+++ src/main/resources/templates/version.html
@@ -5,6 +5,6 @@
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
 </head>
 <body>
-    <p th:text="'Hello e-tax invoice'" />
+    <p th:text="'Version 0.0.1-SNAPSHOT'" />
 </body>
 </html>
