--- src/test/java/th/ac/kmitl/it/group5/etaxinvoice/ETaxInvoiceApplicationTests.java
@@ -1,17 +1,0 @@
-package th.ac.kmitl.it.group5.etaxinvoice;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class ETaxInvoiceApplicationTests {
-
-	@Test
-	public void contextLoads() {
-		// TODO
-	}
-
-}
